+++ advent/Days/2017/Day19.cs
@@ -18,7 +18,7 @@ namespace advent.Days._2017
             return maze.Steps;
         }
 
-        public override int Part1(string input)
+        public override long Part1(string input)
         {
             return Part0(input);
         }
